refactor: centralize card styles
